refactor(loss): break recipes->components import cycle; fix copyright year
Two follow-ups on the latest claude review (PR #2191):

1. **Layering inversion fixed.** ``components/loss/mtp.py``'s
   ``PipelineCausalLMLoss.forward`` had a deferred ``from
   nemo_automodel.recipes.llm.train_ft import calculate_loss,
   calculate_mtp_loss`` to dodge an import-time cycle. The dependency
   direction was wrong: components should never reach into recipes.

   - Extract ``calculate_loss`` (already model-agnostic, used by KD too)
     into a new ``nemo_automodel/components/loss/utils.py``.
   - Move ``calculate_mtp_loss`` alongside ``PipelineCausalLMLoss`` in
     ``components/loss/mtp.py``; it now imports ``calculate_loss`` from
     ``..utils`` at module scope.
   - ``recipes/llm/train_ft.py`` re-imports both names from the new
     locations, preserving the public ``train_ft.calculate_loss`` /
     ``train_ft.calculate_mtp_loss`` import surface that
     ``recipes/llm/kd.py`` and ``tests/functional_tests/checkpoint/
     test_dcp.py`` already rely on. Drop now-unused
     ``get_lm_head_module`` / ``get_lm_head_weight`` imports from
     train_ft.

2. **Copyright year.** ``deepseek_v4/mtp.py`` is a new file added in
   2026, not 2025; align with sibling new file ``loss/mtp.py``.

No behavior change. 179 unit tests across the touched modules pass
locally.
