+++ nemo_automodel/components/loss/mtp.py
@@ -12,11 +12,82 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+from typing import Optional
+
 import torch
 import torch.nn as nn
 
-from nemo_automodel.components.models.common.mtp import get_mtp_loss_scaling_factor
+from nemo_automodel.components.loss.linear_ce import FusedLinearCrossEntropy
+from nemo_automodel.components.loss.utils import calculate_loss
+from nemo_automodel.components.models.common.mtp import get_mtp_loss_scaling_factor, roll_tensor
 from nemo_automodel.components.training.model_output_utils import get_final_hidden_states
+from nemo_automodel.components.utils.model_utils import get_lm_head_module
+
+
+def calculate_mtp_loss(
+    loss_fn,
+    *,
+    mtp_per_depth_h: list[torch.Tensor],
+    labels: torch.Tensor,
+    model: nn.Module,
+    scaling_factor: float = 0.1,
+    num_label_tokens: Optional[int] = None,
+    ignore_index: int = -100,
+) -> torch.Tensor:
+    """Compute the DeepSeek-V3 Multi-Token Prediction auxiliary loss.
+
+    Each depth's CE is dispatched through :func:`calculate_loss` with the
+    same loss class as the main path, so MTP inherits FusedLinearCrossEntropy
+    / MaskedCrossEntropy memory and numerical characteristics.
+
+    Args:
+        loss_fn: Configured per-token loss class (same instance the main
+            path uses).
+        mtp_per_depth_h: Per-depth hidden states from the model's MTP head,
+            one ``[B, S, H]`` tensor per depth.
+        labels: Original (unshifted) labels.
+        model: The wrapped model; used to fetch the shared LM head when the
+            loss class needs materialized logits (non-FusedLinearCE path).
+        scaling_factor: Coefficient applied to the summed per-depth CE.
+        num_label_tokens: Total non-ignore label tokens (forwarded to the
+            base loss for sum-reduction normalization).
+        ignore_index: Label value masked out of the CE loss for the trailing
+            ``k+1`` rolled positions at depth ``k``.
+
+    Returns:
+        Scalar MTP loss with autograd graph.
+    """
+    D = len(mtp_per_depth_h)
+    cur_labels = labels
+    total = mtp_per_depth_h[0].new_zeros(())
+    for k, h_k in enumerate(mtp_per_depth_h):
+        cur_labels = roll_tensor(cur_labels, shifts=-1, dim=-1)
+        masked = cur_labels.clone()
+        n_invalid = min(k + 1, masked.shape[-1])
+        masked[..., -n_invalid:] = ignore_index
+
+        if isinstance(loss_fn, FusedLinearCrossEntropy):
+            depth_loss = calculate_loss(
+                loss_fn,
+                hidden_states=h_k,
+                labels=masked,
+                model=model,
+                num_label_tokens=num_label_tokens,
+            )
+        else:
+            lm_head = get_lm_head_module(model)
+            if lm_head is None:
+                raise ValueError("lm_head module not found in model")
+            depth_loss = calculate_loss(
+                loss_fn,
+                logits=lm_head(h_k),
+                labels=masked,
+                model=model,
+                num_label_tokens=num_label_tokens,
+            )
+        total = total + depth_loss
+
+    return total * (scaling_factor / D)
 
 
 class PipelineCausalLMLoss(nn.Module):
@@ -38,8 +109,6 @@ class PipelineCausalLMLoss(nn.Module):
             hidden_states = get_final_hidden_states(output)
             mtp_per_depth_h = getattr(output, "mtp_per_depth_h", None)
             scaling_factor = getattr(output, "mtp_loss_scaling_factor", get_mtp_loss_scaling_factor(self.model))
-
-        from nemo_automodel.recipes.llm.train_ft import calculate_loss, calculate_mtp_loss
 
         loss = calculate_loss(
             self.loss_fn,
