+++ nemo_automodel/components/loss/utils.py
@@ -1,0 +1,51 @@
+# Copyright (c) 2026, NVIDIA CORPORATION. All rights reserved.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+import torch
+
+from nemo_automodel.components.loss.linear_ce import FusedLinearCrossEntropy
+from nemo_automodel.components.utils.model_utils import get_lm_head_weight
+
+
+def calculate_loss(loss_fn, **kwargs) -> torch.Tensor:
+    """Calculate the loss.
+
+    Args:
+        loss_fn: Loss function.
+        **kwargs: Keyword arguments for the loss function.
+
+    Returns:
+        The loss.
+    """
+    loss_fn_kwargs = {"num_label_tokens": kwargs.pop("num_label_tokens", None)}
+    if isinstance(loss_fn, FusedLinearCrossEntropy):
+        model = kwargs.pop("model")
+        labels = kwargs.pop("labels")
+        lm_head = get_lm_head_weight(model)
+        loss_fn_kwargs.update(
+            {
+                "hidden_states": kwargs.pop("hidden_states"),
+                "labels": labels,
+                "lm_weight": lm_head,
+            }
+        )
+    else:
+        loss_fn_kwargs.update(
+            {
+                "logits": kwargs.pop("logits"),
+                "labels": kwargs.pop("labels"),
+            }
+        )
+
+    return loss_fn(**loss_fn_kwargs)
